Use secondary phase to compute event removing width
Now the weak secondary phase is used to compute the time width for the event
removal. Previously, the calculations were based on duration and period, but
over 1 thousand TCE wolud be removed entirely from the time series. A new
computational strategy ensures that the time series is preserved for any TCE.