change: change convert logic of FitsConverter
Previously, FitsConverter converted FITS time series files to parquet
batch files (multiple time series in one file). This was a convenient
approach because the number of output files was small and could be
searched using DuckDB or other OLAP databases, but reading these files
was very slow (a dozen to tens of seconds). Now all files related to
the same target (via target_id) are written to a single output file
named in the format "{target_id}.json/parquet" and can be read quickly.